Updated main.py
A friend provided me with wonderful suggestions to help make my program much more clear. More of the routines are kept as functions to help with readability.  I also changed how I was importing the functions and settings in order to provide a clear location for the functions and settings used.