Updated settings.py
As part of the suggestions from my friend, I slightly changed the settings.py file. I also added a better way to find the boundaries of the winner panel boxes that allows the size of the boxes to change but not require any new calculations of where one has to click in order to select the winner panel boxes.